requirements.txt
Flask==3.1.2
Flask-Login==0.6.3
SQLAlchemy==2.0.43
python-dotenv==1.1.1
boto3==1.35.95
requests==2.32.3
pytest==8.4.2
